Address the comment from @trustin
