+++ it/mirror/build.gradle
@@ -2,6 +2,6 @@ dependencies {
     testImplementation libs.sshd
     testImplementation libs.jsch
 
-    testImplementation project(':mirror')
+    testImplementation project(':server-mirror-git')
     testImplementation project(':server-auth:shiro')
 }
