--- mirror/src/main/resources/META-INF/services/com.linecorp.centraldogma.server.plugin.Plugin
@@ -1,1 +1,0 @@
-com.linecorp.centraldogma.server.internal.mirror.DefaultMirroringServicePlugin
