--- mirror/src/test/java/com/linecorp/centraldogma/server/internal/mirror/MirroringServicePluginTest.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2020 LINE Corporation
- *
- * LINE Corporation licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package com.linecorp.centraldogma.server.internal.mirror;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
-
-import org.junit.jupiter.api.Test;
-
-import com.linecorp.centraldogma.server.CentralDogmaConfig;
-import com.linecorp.centraldogma.server.internal.PluginGroup;
-import com.linecorp.centraldogma.server.plugin.PluginTarget;
-
-class MirroringServicePluginTest {
-
-    @Test
-    void confirmDefaultMirroringServiceLoadedDependingOnConfig() {
-        final CentralDogmaConfig cfg = mock(CentralDogmaConfig.class);
-        when(cfg.isMirroringEnabled()).thenReturn(true);
-        final PluginGroup group1 = PluginGroup.loadPlugins(PluginTarget.LEADER_ONLY, cfg);
-        assertThat(group1).isNotNull();
-        assertThat(group1.findFirstPlugin(DefaultMirroringServicePlugin.class)).isPresent();
-
-        when(cfg.isMirroringEnabled()).thenReturn(false);
-        final PluginGroup group2 = PluginGroup.loadPlugins(PluginTarget.LEADER_ONLY, cfg);
-        assertThat(group2).isNull();
-    }
-}
