+++ server-mirror-git/src/main/resources/META-INF/services/com.linecorp.centraldogma.server.mirror.MirrorProvider
@@ -1,0 +1,1 @@
+com.linecorp.centraldogma.server.internal.mirror.DefaultMirrorProvider
