+++ server/src/main/java/com/linecorp/centraldogma/server/CentralDogma.java
@@ -115,7 +115,6 @@ import com.linecorp.centraldogma.server.auth.SessionManager;
 import com.linecorp.centraldogma.server.command.Command;
 import com.linecorp.centraldogma.server.command.CommandExecutor;
 import com.linecorp.centraldogma.server.command.StandaloneCommandExecutor;
-import com.linecorp.centraldogma.server.internal.PluginGroup;
 import com.linecorp.centraldogma.server.internal.admin.auth.CachedSessionManager;
 import com.linecorp.centraldogma.server.internal.admin.auth.CsrfTokenAuthorizer;
 import com.linecorp.centraldogma.server.internal.admin.auth.ExpiredSessionDeletingSessionManager;
@@ -136,6 +135,7 @@ import com.linecorp.centraldogma.server.internal.api.WatchService;
 import com.linecorp.centraldogma.server.internal.api.auth.ApplicationTokenAuthorizer;
 import com.linecorp.centraldogma.server.internal.api.converter.HttpApiRequestConverter;
 import com.linecorp.centraldogma.server.internal.api.converter.HttpApiResponseConverter;
+import com.linecorp.centraldogma.server.internal.mirror.DefaultMirroringServicePlugin;
 import com.linecorp.centraldogma.server.internal.replication.ZooKeeperCommandExecutor;
 import com.linecorp.centraldogma.server.internal.storage.project.DefaultProjectManager;
 import com.linecorp.centraldogma.server.internal.storage.project.SafeProjectManager;
@@ -145,7 +145,6 @@ import com.linecorp.centraldogma.server.internal.thrift.CentralDogmaTimeoutSched
 import com.linecorp.centraldogma.server.internal.thrift.TokenlessClientLogger;
 import com.linecorp.centraldogma.server.metadata.MetadataService;
 import com.linecorp.centraldogma.server.metadata.MetadataServiceInjector;
-import com.linecorp.centraldogma.server.mirror.MirroringServicePlugin;
 import com.linecorp.centraldogma.server.plugin.Plugin;
 import com.linecorp.centraldogma.server.plugin.PluginTarget;
 import com.linecorp.centraldogma.server.storage.project.ProjectManager;
@@ -271,8 +270,8 @@ public class CentralDogma implements AutoCloseable {
         if (pluginsForLeaderOnly == null) {
             return Optional.empty();
         }
-        return pluginsForLeaderOnly.findFirstPlugin(MirroringServicePlugin.class)
-                                   .map(MirroringServicePlugin::mirroringService);
+        return pluginsForLeaderOnly.findFirstPlugin(DefaultMirroringServicePlugin.class)
+                                   .map(DefaultMirroringServicePlugin::mirroringService);
     }
 
     /**
