+++ server/src/main/resources/META-INF/services/com.linecorp.centraldogma.server.plugin.Plugin
@@ -1,1 +1,2 @@
+com.linecorp.centraldogma.server.internal.mirror.DefaultMirroringServicePlugin
 com.linecorp.centraldogma.server.internal.storage.PurgeSchedulingServicePlugin
