+++ settings.gradle
@@ -15,10 +15,10 @@ includeWithFlags ':client:java-spring-boot3-autoconfigure', 'java17', 'publish',
 includeWithFlags ':client:java-spring-boot3-starter',       'java17', 'publish', 'relocate'
 includeWithFlags ':common',                                 'java', 'publish', 'shade', 'trim'
 includeWithFlags ':common-legacy',                          'java', 'publish', 'relocate'
-includeWithFlags ':mirror',                                 'java17', 'publish', 'relocate'
 includeWithFlags ':server',                                 'java', 'publish', 'relocate'
 includeWithFlags ':server-auth:saml',                       'java', 'publish', 'relocate'
 includeWithFlags ':server-auth:shiro',                      'java', 'publish', 'relocate'
+includeWithFlags ':server-mirror-git',                      'java11', 'publish', 'relocate'
 includeWithFlags ':testing:testing-common',                 'java', 'publish', 'relocate'
 includeWithFlags ':testing:junit',                          'java', 'publish', 'relocate'
 includeWithFlags ':testing:junit4',                         'java', 'publish', 'relocate'
@@ -26,7 +26,7 @@ includeWithFlags ':testing:junit4',                         'java', 'publish', '
 project(':testing:testing-common').projectDir = file('testing/common')
 
 // Unpublished Java projects
-includeWithFlags ':it:it-mirror',     'java17', 'relocate'
+includeWithFlags ':it:it-mirror',     'java11', 'relocate'
 // Set correct directory names
 project(':it:it-mirror').projectDir = file('it/mirror')
 includeWithFlags ':it:it-server',     'java', 'relocate'
